move starwars test helpers into separate file
